WIP wrap fields
The semantics of default is incorrect:
- In Conformist, the decoder returns it when the field is absent.
- Here, it only dictates the default "value" of the input.

The user may forget to fill an input while we could have prevented
him from doing (using the "required" attribute), and the decoding logic
is wrong: the browser sends an empty string for an empty numerical input.
I might have to map lists before conversion happens.